requirements.txt
PySimpleGUI
pyserial
py2app
python-dotenv
customtkinter